Develop

+++ DAL/DALPhongBan.cs
@@ -12,14 +12,26 @@ namespace DAL
     public class DALPhongBan
     {
         private readonly string connectionString;
-        private readonly PersonnelManagementDataContextDataContext _dbContextPB;
+        private readonly PersonnelManagementDataContextDataContext _dbContext;
 
         public DALPhongBan(string conn)
         {
             connectionString = conn;
-            _dbContextPB = new PersonnelManagementDataContextDataContext(conn);
+            _dbContext = new PersonnelManagementDataContextDataContext(conn);
         }
 
+        public DTOPhongBan FindPhongBanByIdChucVu(int id)
+        {
+            DTOPhongBan phongBan = (from pb in _dbContext.PhongBans
+                                    join cv in _dbContext.ChucVus on pb.id equals cv.idPhongBan
+                                    select new DTOPhongBan
+                                    {
+                                        Id = pb.id,
+                                        TenPhongBan = pb.TenPhongBan,
+                                        MoTa = pb.Mota
+                                    }).FirstOrDefault();
+            return phongBan;
+        }
         public DataTable GetAllPhongBan()
         {
             using (SqlConnection conn = new SqlConnection(connectionString))
@@ -76,6 +88,14 @@ namespace DAL
         {
             using (SqlConnection conn = new SqlConnection(connectionString))
             {
+                var ktraIDPB = _dbContext.ChucVus.FirstOrDefault(cv => cv.idPhongBan == id);
+                var ktraIDNV = _dbContext.NhanViens.FirstOrDefault(cv => cv.idPhongBan == id);
+
+                if (ktraIDPB != null || ktraIDPB != null)
+                {
+                    return false;
+                }
+
                 string query = "DELETE FROM PhongBan WHERE id=@id";
                 SqlCommand cmd = new SqlCommand(query, conn);
                 cmd.Parameters.AddWithValue("@id", id);
@@ -85,9 +105,9 @@ namespace DAL
         }
 
         // Lay ten phong ban
-        public string LayTenPhongBan(int id) => _dbContextPB.PhongBans.Where(p => p.id == id).Select(p => p.TenPhongBan).FirstOrDefault().ToString() ?? string.Empty;
+        public string LayTenPhongBan(int id) => _dbContext.PhongBans.FirstOrDefault(p => p.id == id).TenPhongBan ?? string.Empty;
 
         // Lay ds phong ban
-        public List<PhongBan> LayDsPhongBan() => _dbContextPB.PhongBans.ToList();
+        public List<PhongBan> LayDsPhongBan() => _dbContext.PhongBans.ToList();
     }
 }
